% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_census_pums.R
\name{download_census_pums}
\alias{download_census_pums}
\title{Download U.S. Census Microdata}
\usage{
download_census_pums(dir, state, year = 2021, level = "both",
  one_year = TRUE, calculate_error = FALSE, crosswalk = TRUE,
  geoids = NULL, verbose = TRUE)
}
\arguments{
\item{dir}{Directory in which to save the file(s).}

\item{state}{Postal or FIPS code of the state.}

\item{year}{4 digit year, between 2005 and the most recent year.}

\item{level}{A character indicating whether to get the person- or household-level sample. Defaults to both.}

\item{one_year}{Logical; if \code{FALSE}, will get the 5-year estimates rather than the 1-year file. If
not specified, will fall back to the 5-year file if the 1-year file is not available.}

\item{calculate_error}{Logical; if \code{TRUE}, will calculate standard errors for each variable using
the Successive Difference Replicate Formula from the
[PUMS handbook](https://www.census.gov/programs-surveys/acs/library/handbooks/pums.html).}

\item{crosswalk}{Logical; if \code{FALSE}, will not retrieve the PUMA relationship files for associating
Census tracts with PUM areas. This will be treated as \code{TRUE} if
\code{geoids} is specified.}

\item{geoids}{A vector of county, tract, or block group GEOIDs within the specified state to
select PUM areas by; defaults to all areas. If specified, \code{crosswalk} will be treated as \code{TRUE}.}

\item{verbose}{Logical; if \code{FALSE}, will not print status messages.}
}
\value{
A list with entries for \code{year} and \code{state} (as specified),
\code{dictionary} (containing the data dictionary for that year), \code{household} and/or \code{person}
(with survey data), and optionally \code{household_error} and/or \code{person_error}
(if \code{calculate_error} is \code{TRUE}), and \code{crosswalk} (if \code{crosswalk} is \code{TRUE} or
\code{geoids} is specified).
}
\description{
Download and load U.S. census American Community Survey (ACS) Public Use Microdata Samples (PUMS):
[census.gov/programs-surveys/acs/microdata.html](https://www.census.gov/programs-surveys/acs/microdata.html)
}
\examples{
\dontrun{
download_census_pums(".", "va")
}
}
