% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parent_polygons.R
\name{make_parent_polygons}
\alias{make_parent_polygons}
\title{Make Higher-Order Polygons}
\usage{
make_parent_polygons(polygons, ids = NULL, n = 2, strict_n = TRUE,
  n_as_min = FALSE, buffer_dist = 5e-05, min_overlap = NULL,
  verbose = TRUE)
}
\arguments{
\item{polygons}{An \code{sf} object with polygons to be reformed.}

\item{ids}{A vector of IDs the same length as \code{polygons}, or the column name
in \code{polygons} to use as IDs.}

\item{n}{Number of polygons to aim to combine for each new polygon.}

\item{strict_n}{Logical; if \code{TRUE} (default), \code{n} represents the number of
intersecting polygons to select, depending on availability (with a minimum of 2).
If \code{FALSE}, \code{n} represents the number of nearest polygons (by centroid)
to use when calculating a box to use when selecting polygons (with a minimum of 1).}

\item{n_as_min}{Logical; if \code{TRUE}, will merge any parents with fewer than \code{n}
children with a random neighbor. Otherwise (and by default), parents may have fewer than \code{n}
children. Applies if \code{strict_n} is \code{TRUE}.}

\item{buffer_dist}{Distance around each initial shape or set of shapes, used to define
neighboring shapes. Applies if \code{strict_n} is \code{TRUE}}

\item{min_overlap}{Minimal area of overlap between potential neighboring shapes and
the buffered target shape, used to define neighboring shapes.
Applies if \code{strict_n} is \code{TRUE}}

\item{verbose}{Logical; if \code{FALSE}, will not print status messages.}
}
\value{
A list with entries for \code{new} (an \code{sf} object containing the new polygons)
and \code{map} (a list mapping between the old and new polygon indices).
}
\description{
Randomly combine given polygons into contiguous, larger polygons.
}
