% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_parent_polygons.R
\name{make_parent_polygons}
\alias{make_parent_polygons}
\title{Make Higher-Order Polygons}
\usage{
make_parent_polygons(polygons, n = 3, verbose = TRUE)
}
\arguments{
\item{polygons}{An \code{sf} object with polygons to be reformed.}

\item{n}{Number of polygons to aim to combine for each new polygon.}

\item{verbose}{Logical; if \code{FALSE}, will not print status messages.}
}
\value{
A list with entries for \code{new} (an \code{sf} object containing the new polygons)
and \code{map} (a list mapping between the old and new polygon indices).
}
\description{
Randomly combine given polygons into contiguous, larger polygons.
}
